requirements.txt
python-telegram-bot>=20.0
python-dotenv
firebase-admin>=6.0.0
